+++ lib/application/widgets/material/asset_icon_button.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/material.dart';
+import 'package:layoutr/layoutr.dart';
 
 const double _iconButtonDefaultSize = 24;
-const EdgeInsets _iconButtonDefaultEdgeInsets = EdgeInsets.all(8);
+const Spacing _iconButtonDefaultEdgeInsetsSpacing = Spacing.medium;
 
 /// Creates a new [IconButton] using an [asset].
 class AssetIconButton extends StatelessWidget {
@@ -28,7 +29,7 @@ class AssetIconButton extends StatelessWidget {
     // Even though `ImageIcon` already uses the theme to specify the icon size, we must enforce this through the
     // `IconButton` constructor, otherwise it will be overridden by the default argument.
     final normalizedIconSize = iconSize ?? Theme.of(context).iconTheme.size ?? _iconButtonDefaultSize;
-    final normalizedPadding = padding ?? _iconButtonDefaultEdgeInsets;
+    final normalizedPadding = padding ?? EdgeInsets.all(context.rawSpacing(_iconButtonDefaultEdgeInsetsSpacing));
 
     return IconButton(
       icon: icon,