+++ CHANGELOG.md
@@ -12,6 +12,7 @@ a beta or production release, they must be documented here).
 
 ## Updated
 - Android & iOS Fastfiles to automatically distribute uploaded builds to external testers.
+- `AssetIconButton` default insets padding.
 
 ## Fixed
 - Removed misleading `QuillEditor` cursor in `readOnly` mode.